--- src/ch3-hello-fastapi/main.py
@@ -1,14 +1,0 @@
-import fastapi
-import uvicorn
-
-app = fastapi.FastAPI()
-
-
-@app.get("/")
-def index():
-    return "Hello world"
-
-
-if __name__ == '__main__':
-    uvicorn.run(app)
-
