--- src/ch3-hello-fastapi/requirements.txt
@@ -1,2 +1,0 @@
-fastapi
-uvicorn
