+++ src/viewmodels/home/indexviewmodel.py
@@ -1,0 +1,26 @@
+from starlette.requests import Request
+
+from viewmodels.shared.viewmodel import ViewModelBase
+
+
+class IndexViewModel(ViewModelBase):
+    def __init__(self, request: Request):
+        super().__init__(request)
+
+        self.release_count: int = 1
+        self.user_count: int = 2
+        self.package_count: int = 3
+        self.packages: list = []
+
+    """
+    {
+        'package_count': 274_000,
+        'release_count': 2_234_847,
+        'user_count': 73_874,
+        'packages': [
+            {'id': 'fastapi', 'summary': "A great web framework"},
+            {'id': 'uvicorn', 'summary': "Your favorite ASGI server"},
+            {'id': 'httpx', 'summary': "Requests for an async world"},
+        ]
+    }
+    """
