+++ src/views/home.py
@@ -1,0 +1,20 @@
+import fastapi
+from fastapi_chameleon import template
+from starlette.requests import Request
+
+from viewmodels.home.indexviewmodel import IndexViewModel
+
+router = fastapi.APIRouter()
+
+
+@router.get('/')
+@template()
+def index(request: Request):
+    vm = IndexViewModel(request)
+    return vm.to_dict()
+
+
+@router.get('/about')
+@template()
+def about():
+    return {}
